refactor listings
